kafka_realtime_pipeline/requirements.txt
# Kafka - using maintained fork for Python 3.13 compatibility
kafka-python-ng>=2.2.0

# Data generation
faker>=30.0.0
numpy>=1.26.0

# Database
psycopg2-binary>=2.9.9

# Data processing and visualization
pandas>=2.1.0
plotly>=5.18.0
streamlit>=1.29.0
sqlalchemy>=2.0.23

# Machine Learning for anomaly detection
scikit-learn>=1.3.2

# Flink (optional - for full Flink integration)
apache-flink==1.18.0
# Note: Full PyFlink requires Java 11+ and additional setup
# For simplified demo, we use SQL-based aggregation